Created a new hook mean't to handle the scroll event of coinTable. Stil WIP (Work in progress)
